requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
attrs==21.4.0
azure-core==1.26.3
azure-storage-blob==12.14.1
blessed==1.19.1
bpython==0.23
cattrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
curtsies==0.4.1
cwcwidth==0.1.8
exceptiongroup==1.1.0
gooddata-afm-client==1.2.0
gooddata-api-client==1.2.0
gooddata-metadata-client==1.2.0
gooddata-scan-client==1.2.0
gooddata-sdk==1.2.0
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pyxdg==0.28
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
zipp==3.11.0